fix: place definitions at root
