+++ tests/Generator/Fixtures/NestedTypedArrayProperty/Output/Record.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Ns\NestedTypedArrayProperty\Definitions;
+namespace Ns\NestedTypedArrayProperty;
 
 class Record
 {
@@ -67,25 +67,46 @@ class Record
                 'maxItems' => 1,
             ],
         ],
+        'definitions' => [
+            'Phone' => [
+                'type' => 'object',
+                'properties' => [
+                    'foo' => [
+                        'type' => 'string',
+                    ],
+                ],
+            ],
+            'Fio' => [
+                'type' => 'object',
+                'properties' => [
+                    'bar' => [
+                        'type' => [
+                            'null',
+                            'string',
+                        ],
+                    ],
+                ],
+            ],
+        ],
     ];
 
     /**
-     * @var Ns\NestedTypedArrayProperty\Definitions\Phone[]|null
+     * @var Phone[]|null
      */
     private ?array $dataArray = null;
 
     /**
-     * @var Ns\NestedTypedArrayProperty\Definitions\Phone[][]|null
+     * @var Phone[][]|null
      */
     private ?array $dataArrayNested = null;
 
     /**
-     * @var (Ns\NestedTypedArrayProperty\Definitions\Phone|Ns\NestedTypedArrayProperty\DefinitionsFio)[]|null
+     * @var (Phone|Fio)[]|null
      */
     private ?array $dataArrayAnyOf = null;
 
     /**
-     * @var ((Ns\NestedTypedArrayProperty\Definitions\Phone|Ns\NestedTypedArrayProperty\DefinitionsFio)[])[]|null
+     * @var ((Phone|Fio)[])[]|null
      */
     private ?array $dataArrayNestedAnyOf = null;
 
@@ -97,7 +118,7 @@ class Record
     }
 
     /**
-     * @return Ns\NestedTypedArrayProperty\Definitions\Phone[]|null
+     * @return Phone[]|null
      */
     public function getDataArray() : ?array
     {
@@ -105,7 +126,7 @@ class Record
     }
 
     /**
-     * @return Ns\NestedTypedArrayProperty\Definitions\Phone[][]|null
+     * @return Phone[][]|null
      */
     public function getDataArrayNested() : ?array
     {
@@ -113,7 +134,7 @@ class Record
     }
 
     /**
-     * @return (Ns\NestedTypedArrayProperty\Definitions\Phone|Ns\NestedTypedArrayProperty\DefinitionsFio)[]|null
+     * @return (Phone|Fio)[]|null
      */
     public function getDataArrayAnyOf() : ?array
     {
@@ -121,7 +142,7 @@ class Record
     }
 
     /**
-     * @return ((Ns\NestedTypedArrayProperty\Definitions\Phone|Ns\NestedTypedArrayProperty\DefinitionsFio)[])[]|null
+     * @return ((Phone|Fio)[])[]|null
      */
     public function getDataArrayNestedAnyOf() : ?array
     {
@@ -129,7 +150,7 @@ class Record
     }
 
     /**
-     * @param Ns\NestedTypedArrayProperty\Definitions\Phone[] $dataArray
+     * @param Phone[] $dataArray
      * @return self
      */
     public function withDataArray(array $dataArray) : self
@@ -158,7 +179,7 @@ class Record
     }
 
     /**
-     * @param Ns\NestedTypedArrayProperty\Definitions\Phone[][] $dataArrayNested
+     * @param Phone[][] $dataArrayNested
      * @return self
      */
     public function withDataArrayNested(array $dataArrayNested) : self
@@ -187,7 +208,7 @@ class Record
     }
 
     /**
-     * @param (Ns\NestedTypedArrayProperty\Definitions\Phone|Ns\NestedTypedArrayProperty\DefinitionsFio)[] $dataArrayAnyOf
+     * @param (Phone|Fio)[] $dataArrayAnyOf
      * @return self
      */
     public function withDataArrayAnyOf(array $dataArrayAnyOf) : self
@@ -216,7 +237,7 @@ class Record
     }
 
     /**
-     * @param ((Ns\NestedTypedArrayProperty\Definitions\Phone|Ns\NestedTypedArrayProperty\DefinitionsFio)[])[] $dataArrayNestedAnyOf
+     * @param ((Phone|Fio)[])[] $dataArrayNestedAnyOf
      * @return self
      */
     public function withDataArrayNestedAnyOf(array $dataArrayNestedAnyOf) : self
@@ -265,17 +286,17 @@ class Record
             static::validateInput($input);
         }
 
-        $dataArray = isset($input->{'dataArray'}) ? array_map(fn(array|object $i): Ns\NestedTypedArrayProperty\DefinitionsPhone => Ns\NestedTypedArrayProperty\Definitions\Phone::buildFromInput($i, $validate), $input->{'dataArray'}) : null;
-        $dataArrayNested = isset($input->{'dataArrayNested'}) ? array_map(fn($i) => array_map(fn(array|object $i): Ns\NestedTypedArrayProperty\DefinitionsPhone => Ns\NestedTypedArrayProperty\Definitions\Phone::buildFromInput($i, $validate), $i), $input->{'dataArrayNested'}) : null;
+        $dataArray = isset($input->{'dataArray'}) ? array_map(fn(array|object $i): Phone => Phone::buildFromInput($i, $validate), $input->{'dataArray'}) : null;
+        $dataArrayNested = isset($input->{'dataArrayNested'}) ? array_map(fn($i) => array_map(fn(array|object $i): Phone => Phone::buildFromInput($i, $validate), $i), $input->{'dataArrayNested'}) : null;
         $dataArrayAnyOf = isset($input->{'dataArrayAnyOf'}) ? array_map(fn($i) => match (true) {
             default => null,
-            Ns\NestedTypedArrayProperty\Definitions\Phone::validateInput($i, true) => Ns\NestedTypedArrayProperty\Definitions\Phone::buildFromInput($i, $validate),
-            Ns\NestedTypedArrayProperty\Definitions\Fio::validateInput($i, true) => Ns\NestedTypedArrayProperty\Definitions\Fio::buildFromInput($i, $validate),
+            Phone::validateInput($i, true) => Phone::buildFromInput($i, $validate),
+            Fio::validateInput($i, true) => Fio::buildFromInput($i, $validate),
         }, $input->{'dataArrayAnyOf'}) : null;
         $dataArrayNestedAnyOf = isset($input->{'dataArrayNestedAnyOf'}) ? array_map(fn($i) => array_map(fn($i) => match (true) {
             default => null,
-            Ns\NestedTypedArrayProperty\Definitions\Phone::validateInput($i, true) => Ns\NestedTypedArrayProperty\Definitions\Phone::buildFromInput($i, $validate),
-            Ns\NestedTypedArrayProperty\Definitions\Fio::validateInput($i, true) => Ns\NestedTypedArrayProperty\Definitions\Fio::buildFromInput($i, $validate),
+            Phone::validateInput($i, true) => Phone::buildFromInput($i, $validate),
+            Fio::validateInput($i, true) => Fio::buildFromInput($i, $validate),
         }, $i), $input->{'dataArrayNestedAnyOf'}) : null;
 
         $obj = new self();
@@ -295,21 +316,21 @@ class Record
     {
         $output = [];
         if (isset($this->dataArray)) {
-            $output['dataArray'] = array_map(fn(Ns\NestedTypedArrayProperty\DefinitionsPhone $i): array => $i->toJson(), $this->dataArray);
+            $output['dataArray'] = array_map(fn(Phone $i): array => $i->toJson(), $this->dataArray);
         }
         if (isset($this->dataArrayNested)) {
-            $output['dataArrayNested'] = array_map(fn($i) => array_map(fn(Ns\NestedTypedArrayProperty\DefinitionsPhone $i): array => $i->toJson(), $i), $this->dataArrayNested);
+            $output['dataArrayNested'] = array_map(fn($i) => array_map(fn(Phone $i): array => $i->toJson(), $i), $this->dataArrayNested);
         }
         if (isset($this->dataArrayAnyOf)) {
             $output['dataArrayAnyOf'] = array_map(fn($i) => match (true) {
                 default => null,
-                ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsPhone, ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsFio => $i->toJson(),
+                ($i) instanceof Phone, ($i) instanceof Fio => $i->toJson(),
             }, $this->dataArrayAnyOf);
         }
         if (isset($this->dataArrayNestedAnyOf)) {
             $output['dataArrayNestedAnyOf'] = array_map(fn($i) => array_map(fn($i) => match (true) {
                 default => null,
-                ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsPhone, ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsFio => $i->toJson(),
+                ($i) instanceof Phone, ($i) instanceof Fio => $i->toJson(),
             }, $i), $this->dataArrayNestedAnyOf);
         }
 
@@ -347,12 +368,12 @@ class Record
         }
         if (isset($this->dataArrayAnyOf)) {
             $this->dataArrayAnyOf = array_map(fn($i) => match (true) {
-                ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsPhone, ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsFio => $i,
+                ($i) instanceof Phone, ($i) instanceof Fio => $i,
             }, $this->dataArrayAnyOf);
         }
         if (isset($this->dataArrayNestedAnyOf)) {
             $this->dataArrayNestedAnyOf = array_map(fn($i) => array_map(fn($i) => match (true) {
-                ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsPhone, ($i) instanceof Ns\NestedTypedArrayProperty\DefinitionsFio => $i,
+                ($i) instanceof Phone, ($i) instanceof Fio => $i,
             }, $i), $this->dataArrayNestedAnyOf);
         }
     }
